Fix regex and change expected path of the log files in process_client_logs.sh
Fixes regex to capture hour of recorded times in logs and to capture double spaces when the day of the month is a single digit.
Also changes process_client_logs.sh to look in ./var/log for the log files, which, based on the tests, appears to be the expected behavior.

Co-Authored by: @day00090 Kyle Day