Add create_country_dist.sh
This commit adds the script create_country_dist.sh. Currently, it is non-operational, but, were it to be operational, it would function similary to create_username_dist.sh and create_hours_dist.sh. The main differences are that IP addresses are used to find the countries, and Google Geo Chart is used for creating the distribution chart.

Co-authored-by: Erik Rauer rauer007@morris.umn.edu